requirements.txt
streamlit
pandas
openpyxl
pdf2image
pytesseract
pillow
pywin32
